Move loading responsibility for userGreeting to main.js
